requirements.txt
albumentations==1.4.11
einops==0.8.0
ftfy==6.1.1
numpy==1.26.4
omegaconf==2.3.0
opencv_python==4.6.0.66
Pillow==9.2.0
regex==2022.8.17
setuptools==63.4.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.0
